feat: 하단 버튼, 텍스트 추가
