+++ App.tsx
@@ -1,21 +1,2 @@
-import { StatusBar } from "expo-status-bar";
-import { StyleSheet, Text, View } from "react-native";
-import Onboarding from "./components/Onboarding";
-
-export default function App() {
-  return (
-    <View style={styles.container}>
-      <Onboarding />
-      <StatusBar style="auto" />
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-});
+import App from "./src/App";
+export default App;
