+++ src/components/OnboardingItem.tsx
